export_scripts/requirements.txt
# Requirements for ACGI Contact Export Script
requests>=2.28.0
python-dotenv>=0.19.0
